requirements.txt
Flask==3.0.3
gunicorn==21.2.0
#boto3==1.34.162
python-dotenv==1.0.1

#langchain==1.2.1
#langchain-core==1.2.6
#langchain-community==0.4.1
#langchain-aws==0.1.18
